+++ R/mirai-package.R
@@ -103,8 +103,7 @@
     single_url = "only one 'url' should be specified",
     sync_timeout = "initial sync with dispatcher/daemon timed out after 10s",
     url_spec = "numeric value for 'url' is out of bounds",
-    wrong_dots = "'...' arguments should only be of integer, numeric or logical type",
-    xlen_ylen = "length of '.y' must be the length of '.x' or one"
+    wrong_dots = "'...' arguments should only be of integer, numeric or logical type"
   ),
   hash = TRUE
 )